# api/handler/teacher.go
// Copyright 2025 Boyuan-IT-Club
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package handler

import (
	"github.com/Boyuan-IT-Club/Meowpick-Backend/application/dto"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/infra/util/token"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/provider"
	"github.com/Boyuan-IT-Club/Meowpick-Backend/types/consts"
	"github.com/gin-gonic/gin"
)

// CreateTeacher godoc
// @Summary 新建教师
// @Description 新建教师
// @Tags teacher
// @Accept json
// @Produce json
// @Param body body dto.CreateTeacherReq true "CreateTeacherReq"
// @Success 200 {object} Response[dto.CreateTeacherResp]
// @Security Bearer
// @Router /api/teacher/add [post]
func CreateTeacher(c *gin.Context) {
	var req *dto.CreateTeacherReq
	var resp *dto.CreateTeacherResp
	var err error

	if err = c.ShouldBind(&req); err != nil {
		PostProcess(c, req, resp, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().TeacherService.CreateTeacher(c, req)
	PostProcess(c, req, resp, err)
}

// GetTeacherSuggestions godoc
// @Summary 获取教师搜索建议
// @Description 根据关键词获取教师搜索建议
// @Tags teacher
// @Produce json
// @Param keyword query string true "搜索关键词"
// @Success 200 {object} Response[dto.GetTeacherSuggestionsResp]
// @Security Bearer
// @Router /api/teacher/suggest [get]
func GetTeacherSuggestions(c *gin.Context) {
	var req *dto.GetTeacherSuggestionsReq
	var resp *dto.GetTeacherSuggestionsResp
	var err error

	if err = c.ShouldBindQuery(&req); err != nil {
		PostProcess(c, req, resp, err)
		return
	}
	c.Set(consts.CtxUserID, token.GetUserID(c))

	resp, err = provider.Get().TeacherService.GetTeacherSuggestions(c, req)
	PostProcess(c, req, resp, err)
}