# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "dto.CreateTeacherReq": {
                "properties": {
                    "department": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "required": [
                    "department",
                    "name",
                    "title"
                ],
                "type": "object"
            },
            "dto.CreateTeacherResp": {
                "description": "实际业务数据",
                "properties": {
                    "department": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.GetSearchHistoriesResp": {
                "description": "实际业务数据",
                "properties": {
                    "histories": {
                        "items": {
                            "$ref": "#/components/schemas/dto.SearchHistoryVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "dto.GetSearchSuggestionsResp": {
                "description": "实际业务数据",
                "properties": {
                    "suggestions": {
                        "items": {
                            "$ref": "#/components/schemas/dto.SearchSuggestionsVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "dto.GetTeacherSuggestionsResp": {
                "description": "实际业务数据",
                "properties": {
                    "teachers": {
                        "items": {
                            "$ref": "#/components/schemas/dto.TeacherVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "dto.SearchHistoryVO": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "query": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "dto.SearchSuggestionsVO": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.TeacherVO": {
                "properties": {
                    "department": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_CreateTeacherResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.CreateTeacherResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_GetSearchHistoriesResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetSearchHistoriesResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_GetSearchSuggestionsResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetSearchSuggestionsResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_GetTeacherSuggestionsResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetTeacherSuggestionsResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "选课猫后端接口文档",
        "title": "Meowpick Backend API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/search/recent": {
            "get": {
                "description": "获取最近搜索历史",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetSearchHistoriesResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取最近搜索历史",
                "tags": [
                    "search"
                ]
            }
        },
        "/api/search/suggest": {
            "get": {
                "description": "根据关键词获取搜索建议\n根据 type 不同执行不同的搜索逻辑：\n- course：模糊分页搜索课程\n- teacher：精确分页搜索教师开设的课程\n- category：精确分页搜索该类别下的课程\n- department：精确分页搜索该开课院系下的课程",
                "parameters": [
                    {
                        "description": "搜索关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetSearchSuggestionsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取搜索建议",
                "tags": [
                    "search"
                ]
            }
        },
...
        "/api/teacher/add": {
            "post": {
                "description": "新建教师",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.CreateTeacherReq",
                                        "summary": "body",
                                        "description": "CreateTeacherReq"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "CreateTeacherReq",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_CreateTeacherResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "新建教师",
                "tags": [
                    "teacher"
                ]
            }
        },
        "/api/teacher/suggest": {
            "get": {
                "description": "根据关键词获取教师搜索建议",
                "parameters": [
                    {
                        "description": "搜索关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetTeacherSuggestionsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取教师搜索建议",
                "tags": [
                    "teacher"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}